Initialize metrics addon agent
This commit initializes a metrics addon agent which is actually a
Prometheus Agent deployment, that federates metrics from openshift-monitoring
Prometheus and remote-writes them to ACM Hub Thanos.

In general, we do not expect end user to modify this configuration in
any way as we only plan to offer a single metrics sink of ACM Thanos. But
this can change in the future.

Still quite new to addons, so would appreciate any early feedback, and if
I'm doing the correct thing here. :)

TODOs: Add tests, actually run this on a cluster and add facilities for
rules and collectrules to fully replace metrics-collector

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>